backend-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiosqlite==0.19.0
python-dotenv==1.0.0
bcrypt==4.0.1
redis==5.0.1
httpx==0.25.2